previous_sprints/postgres/requirements.txt
psycopg2-binary==2.9.7
django==4.0.4
python-dotenv==0.21.0
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
uwsgi==2.0.21
faker==22.4.0
redis==5.0.1
pydantic==1.10.12